Create DatabaseModule
DatabseModule will be used to provide dependencies for database classes or data stores. In this case it will provide dependencies
for the db helper to use. This firebase injection will ensure that we do not inject any reference to a database in our view
but instead use it in the dbhelper implementation to allow for easier interchanging between the data stores we use, in the event that happens